update to sample : use float value for currency

+++ trunk/samples/showcase/src/com/smartgwt/sample/showcase/client/miniapp/ItemListGrid.java
@@ -39,7 +39,7 @@ public class ItemListGrid extends ListGrid {
                 if (value == null) return null;
                 try {
                     NumberFormat nf = NumberFormat.getFormat("##0.00");
-                    return "$" + nf.format(((Number) value).longValue());
+                    return "$" + nf.format(((Number) value).floatValue());
                 } catch (Exception e) {
                     return value.toString();
                 }
